# inesdata-map-editor-frontend/src/openapi/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "Service responsible for saving a data source mapping and generating the corresponding R2RML / RML mapping file",
    "title": "com.inesdata-map:mapper-backend",
    "version": "1.1.0-SNAPSHOT"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/executions/{id}/files": {
      "get": {
        "operationId": "downloadFile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Download the specified file from the given execution",
        "tags": [
          "Execution"
        ]
      }
    },
    "/file-sources": {
      "get": {
        "operationId": "getFileSourceByType",
        "parameters": [
          {
            "description": "File source type",
            "in": "query",
            "name": "type",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileSourceDTO"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Gets all file sources filtered by type",
        "tags": [
          "FileSource"
        ]
      },
      "post": {
        "operationId": "createFileSource",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "body": {
                    "$ref": "#/components/schemas/FileSourceDTO"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "body"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSourceDTO"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Create given file source",
        "tags": [
          "FileSource"
        ]
      }
    },
    "/file-sources/{id}": {
      "get": {
        "operationId": "getFileSource",
        "parameters": [
          {
            "description": "File source identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileSourceDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Gets given data file source",
        "tags": [
          "FileSource"
        ]
      },
      "put": {
        "operationId": "updateFileSource",
        "parameters": [
          {
            "description": "File source identifier to update",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileSourceDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSourceDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Updates given file source",
        "tags": [
          "FileSource"
        ]
      }
    },
    "/file-sources/{id}/attributes": {
      "get": {
        "operationId": "getFileAttributes",
        "parameters": [
          {
            "description": "File source identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Gets all XML file attributes",
        "tags": [
          "FileSource"
        ]
      }
    },
    "/file-sources/{id}/fields": {
      "get": {
        "operationId": "getFileFields",
        "parameters": [
          {
            "description": "File source identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Gets all CSV file fields",
        "tags": [
          "FileSource"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataSourceDTO": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "DATABASE",
              "FILE"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "FileSourceDTO": {
        "required": [
          "fileType",
          "name"
        ],
        "type": "object",
        "properties": {
          "fileType": {
            "type": "string",
            "enum": [
              "CSV",
              "JSON",
              "XML"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "DATABASE",
              "FILE"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "description": "The file source to save"
      },
...
    }
  }
}